🐞🔁 Engine/Core & Graphics: Template UniformBufferManagement on Blob Type to implement correct per-material behaviour
Since Materials used the same one and only UniformBufferManagement with a DirtyBlob inside, and since UniformBufferManagement::UploadAll() works with a dirty blob, coupled with the fact that there is only 1 UBO for each uniform BLOCK, Different materials having the same shader can not have per-material uniforms! This is because when an attempt is made to upload a particular Material's blob info to the corresponding UBO, the actual Upload() call on the Uniform Buffer is only made when the Material is dirty.

So to fix this issue, UniformBufferManagement is template on the kind of Blob it takes. This way, Intrinsic/Global shader uniform blocks can utilize DirtyBlobs, while the "regular" Materials use a UniformBufferManagement with a regular Blob (without dirty section tracking). This means that for every Material with the same shader, the UBO is set/uploaded before the respective draw call.

Maybe not the most ideal performance wise, but correct from a usage standpoint: Every UNIQUE material should be able to affect the objects it is put on.

Future work: Along with instance uniforms (which may use a separate UBO, or the very same UBO, and indexed with a default block uniform in both scenarios),
arrays of structs can be used inside a UBO for a single Material type, to minimize UBO binding calls (i.e., all Phong materials may live inside a single UBO, which is how the current __Intrinsic_Lighting UBO works).